Step two: restructure chat IPC
